+++ pkg/api/query/grpc.go
@@ -7,14 +7,16 @@ import (
 	"context"
 	"time"
 
-	"github.com/prometheus/prometheus/promql"
-	"github.com/prometheus/prometheus/storage"
-	"github.com/thanos-io/promql-engine/engine"
-	"github.com/thanos-io/promql-engine/logicalplan"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/storage"
+
+	"github.com/thanos-io/promql-engine/api"
+	"github.com/thanos-io/promql-engine/engine"
+	"github.com/thanos-io/promql-engine/logicalplan"
 	"github.com/thanos-io/thanos/pkg/api/query/querypb"
 	"github.com/thanos-io/thanos/pkg/query"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
@@ -26,7 +28,8 @@ type GRPCAPI struct {
 	now                         func() time.Time
 	replicaLabels               []string
 	queryableCreate             query.QueryableCreator
-	engineFactory               *QueryEngineFactory
+	remoteEndpointsCreate       query.RemoteEndpointsCreator
+	queryCreator                queryCreator
 	defaultEngine               querypb.EngineType
 	lookbackDeltaCreate         func(int64) time.Duration
 	defaultMaxResolutionSeconds time.Duration
@@ -35,8 +38,9 @@ type GRPCAPI struct {
 func NewGRPCAPI(
 	now func() time.Time,
 	replicaLabels []string,
-	creator query.QueryableCreator,
-	engineFactory *QueryEngineFactory,
+	queryableCreator query.QueryableCreator,
+	remoteEndpointsCreator query.RemoteEndpointsCreator,
+	queryCreator queryCreator,
 	defaultEngine querypb.EngineType,
 	lookbackDeltaCreate func(int64) time.Duration,
 	defaultMaxResolutionSeconds time.Duration,
@@ -44,8 +48,9 @@ func NewGRPCAPI(
 	return &GRPCAPI{
 		now:                         now,
 		replicaLabels:               replicaLabels,
-		queryableCreate:             creator,
-		engineFactory:               engineFactory,
+		queryableCreate:             queryableCreator,
+		remoteEndpointsCreate:       remoteEndpointsCreator,
+		queryCreator:                queryCreator,
 		defaultEngine:               defaultEngine,
 		lookbackDeltaCreate:         lookbackDeltaCreate,
 		defaultMaxResolutionSeconds: defaultMaxResolutionSeconds,
@@ -94,10 +99,15 @@ func (g *GRPCAPI) Query(request *querypb.QueryRequest, server querypb.Query_Quer
 		query.NoopSeriesStatsReporter,
 	)
 
+	remoteEndpoints := g.remoteEndpointsCreate(
+		replicaLabels,
+		request.EnablePartialResponse,
+	)
+
 	var qry promql.Query
 	if err := tracing.DoInSpanWithErr(ctx, "instant_query_create", func(ctx context.Context) error {
 		var err error
-		qry, err = g.getQueryForEngine(ctx, request, queryable, maxResolution)
+		qry, err = g.getInstantQueryForEngine(ctx, request, queryable, remoteEndpoints, maxResolution)
 		return err
 	}); err != nil {
 		return err
@@ -105,10 +115,16 @@ func (g *GRPCAPI) Query(request *querypb.QueryRequest, server querypb.Query_Quer
 	defer qry.Close()
 
 	var result *promql.Result
-	tracing.DoInSpan(ctx, "range_query_exec", func(ctx context.Context) {
+	tracing.DoInSpan(ctx, "instant_query_exec", func(ctx context.Context) {
 		result = qry.Exec(ctx)
 	})
 	if result.Err != nil {
+		if request.EnablePartialResponse {
+			if err := server.Send(querypb.NewQueryWarningsResponse(err)); err != nil {
+				return err
+			}
+			return nil
+		}
 		return status.Error(codes.Aborted, result.Err.Error())
 	}
 
@@ -146,42 +162,6 @@ func (g *GRPCAPI) Query(request *querypb.QueryRequest, server querypb.Query_Quer
 	return nil
 }
 
-func (g *GRPCAPI) getQueryForEngine(ctx context.Context, request *querypb.QueryRequest, queryable storage.Queryable, maxResolution int64) (promql.Query, error) {
-	lookbackDelta := g.lookbackDeltaCreate(maxResolution * 1000)
-	if request.LookbackDeltaSeconds > 0 {
-		lookbackDelta = time.Duration(request.LookbackDeltaSeconds) * time.Second
-	}
-	engineParam := request.Engine
-	if engineParam == querypb.EngineType_default {
-		engineParam = g.defaultEngine
-	}
-
-	var ts time.Time
-	if request.TimeSeconds == 0 {
-		ts = g.now()
-	} else {
-		ts = time.Unix(request.TimeSeconds, 0)
-	}
-	opts := &engine.QueryOpts{
-		LookbackDeltaParam:     lookbackDelta,
-		EnablePartialResponses: request.EnablePartialResponse,
-	}
-	switch engineParam {
-	case querypb.EngineType_prometheus:
-		queryEngine := g.engineFactory.GetPrometheusEngine()
-		return queryEngine.MakeInstantQuery(ctx, queryable, opts, request.Query, ts)
-	case querypb.EngineType_thanos:
-		queryEngine := g.engineFactory.GetThanosEngine()
-		plan, err := logicalplan.Unmarshal(request.QueryPlan.GetJson())
-		if err != nil {
-			return queryEngine.MakeInstantQuery(ctx, queryable, opts, request.Query, ts)
-		}
-		return queryEngine.MakeInstantQueryFromPlan(ctx, queryable, opts, plan, ts)
-	default:
-		return nil, status.Error(codes.InvalidArgument, "invalid engine parameter")
-	}
-}
-
 func (g *GRPCAPI) QueryRange(request *querypb.QueryRangeRequest, srv querypb.Query_QueryRangeServer) error {
 	ctx := srv.Context()
 	if request.TimeoutSeconds != 0 {
@@ -217,10 +197,15 @@ func (g *GRPCAPI) QueryRange(request *querypb.QueryRangeRequest, srv querypb.Que
 		query.NoopSeriesStatsReporter,
 	)
 
+	remoteEndpoints := g.remoteEndpointsCreate(
+		replicaLabels,
+		request.EnablePartialResponse,
+	)
+
 	var qry promql.Query
 	if err := tracing.DoInSpanWithErr(ctx, "range_query_create", func(ctx context.Context) error {
 		var err error
-		qry, err = g.getRangeQueryForEngine(ctx, request, queryable)
+		qry, err = g.getRangeQueryForEngine(ctx, request, queryable, remoteEndpoints, maxResolution)
 		return err
 	}); err != nil {
 		return err
@@ -295,45 +280,90 @@ func extractQueryStats(qry promql.Query) *querypb.QueryStats {
 	return stats
 }
 
+func (g *GRPCAPI) getInstantQueryForEngine(
+	ctx context.Context,
+	request *querypb.QueryRequest,
+	queryable storage.Queryable,
+	remoteEndpoints api.RemoteEndpoints,
+	maxResolution int64,
+) (promql.Query, error) {
+	lookbackDelta := g.lookbackDeltaCreate(maxResolution * 1000)
+	if request.LookbackDeltaSeconds > 0 {
+		lookbackDelta = time.Duration(request.LookbackDeltaSeconds) * time.Second
+	}
+	engineParam := request.Engine
+	if engineParam == querypb.EngineType_default {
+		engineParam = g.defaultEngine
+	}
+
+	var ts time.Time
+	if request.TimeSeconds == 0 {
+		ts = g.now()
+	} else {
+		ts = time.Unix(request.TimeSeconds, 0)
+	}
+	opts := &engine.QueryOpts{
+		LookbackDeltaParam: lookbackDelta,
+	}
+
+	var engineType PromqlEngineType
+	switch engineParam {
+	case querypb.EngineType_prometheus:
+		engineType = PromqlEnginePrometheus
+	case querypb.EngineType_thanos:
+		engineType = PromqlEngineThanos
+	default:
+		return nil, status.Error(codes.InvalidArgument, "invalid engine parameter")
+	}
+
+	var qry planOrQuery
+	if plan, err := logicalplan.Unmarshal(request.QueryPlan.GetJson()); err != nil {
+		qry = planOrQuery{plan: plan, query: request.Query}
+	} else {
+		qry = planOrQuery{query: request.Query}
+	}
+	return g.queryCreator.makeInstantQuery(ctx, engineType, queryable, remoteEndpoints, qry, opts, ts)
+}
+
 func (g *GRPCAPI) getRangeQueryForEngine(
 	ctx context.Context,
 	request *querypb.QueryRangeRequest,
 	queryable storage.Queryable,
+	remoteEndpoints api.RemoteEndpoints,
+	maxResolution int64,
 ) (promql.Query, error) {
-	startTime := time.Unix(request.StartTimeSeconds, 0)
-	endTime := time.Unix(request.EndTimeSeconds, 0)
-	interval := time.Duration(request.IntervalSeconds) * time.Second
+	start := time.Unix(request.StartTimeSeconds, 0)
+	end := time.Unix(request.EndTimeSeconds, 0)
+	step := time.Duration(request.IntervalSeconds) * time.Second
 
 	engineParam := request.Engine
 	if engineParam == querypb.EngineType_default {
 		engineParam = g.defaultEngine
 	}
 
-	maxResolution := request.MaxResolutionSeconds
-	if request.MaxResolutionSeconds == 0 {
-		maxResolution = g.defaultMaxResolutionSeconds.Milliseconds() / 1000
-	}
 	lookbackDelta := g.lookbackDeltaCreate(maxResolution * 1000)
 	if request.LookbackDeltaSeconds > 0 {
 		lookbackDelta = time.Duration(request.LookbackDeltaSeconds) * time.Second
 	}
 	opts := &engine.QueryOpts{
-		LookbackDeltaParam:     lookbackDelta,
-		EnablePartialResponses: request.EnablePartialResponse,
+		LookbackDeltaParam: lookbackDelta,
 	}
 
+	var engineType PromqlEngineType
 	switch engineParam {
 	case querypb.EngineType_prometheus:
-		queryEngine := g.engineFactory.GetPrometheusEngine()
-		return queryEngine.MakeRangeQuery(ctx, queryable, opts, request.Query, startTime, endTime, interval)
+		engineType = PromqlEnginePrometheus
 	case querypb.EngineType_thanos:
-		thanosEngine := g.engineFactory.GetThanosEngine()
-		plan, err := logicalplan.Unmarshal(request.QueryPlan.GetJson())
-		if err != nil {
-			return thanosEngine.MakeRangeQuery(ctx, queryable, opts, request.Query, startTime, endTime, interval)
-		}
-		return thanosEngine.MakeRangeQueryFromPlan(ctx, queryable, opts, plan, startTime, endTime, interval)
+		engineType = PromqlEngineThanos
 	default:
 		return nil, status.Error(codes.InvalidArgument, "invalid engine parameter")
 	}
+
+	var qry planOrQuery
+	if plan, err := logicalplan.Unmarshal(request.QueryPlan.GetJson()); err != nil {
+		qry = planOrQuery{plan: plan, query: request.Query}
+	} else {
+		qry = planOrQuery{query: request.Query}
+	}
+	return g.queryCreator.makeRangeQuery(ctx, engineType, queryable, remoteEndpoints, qry, opts, start, end, step)
 }