+++ pkg/store/bucket_test.go
@@ -685,7 +685,7 @@ func TestBucketStore_TSDBInfo(t *testing.T) {
 		{mint: 0, maxt: 1000, extLabels: labels.FromStrings("a", "d")},
 		{mint: 2000, maxt: 3000, extLabels: labels.FromStrings("a", "d")},
 	} {
-		id1, err := e2eutil.CreateBlock(ctx, dir, series, 10, tt.mint, tt.maxt, tt.extLabels, 0, metadata.NoneFunc)
+		id1, err := e2eutil.CreateBlock(ctx, dir, series, 10, tt.mint, tt.maxt, tt.extLabels, 0, metadata.NoneFunc, nil)
 		testutil.Ok(t, err)
 		testutil.Ok(t, block.Upload(ctx, logger, bkt, filepath.Join(dir, id1.String()), metadata.NoneFunc))
 	}
@@ -788,19 +788,19 @@ func TestBucketStore_Sharding(t *testing.T) {
 	bkt := objstore.NewInMemBucket()
 	series := []labels.Labels{labels.FromStrings("a", "1", "b", "1")}
 
-	id1, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "a", "region", "r1"), 0, metadata.NoneFunc)
+	id1, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "a", "region", "r1"), 0, metadata.NoneFunc, nil)
 	testutil.Ok(t, err)
 	testutil.Ok(t, block.Upload(ctx, logger, bkt, filepath.Join(dir, id1.String()), metadata.NoneFunc))
 
-	id2, err := e2eutil.CreateBlock(ctx, dir, series, 10, 1000, 2000, labels.FromStrings("cluster", "a", "region", "r1"), 0, metadata.NoneFunc)
+	id2, err := e2eutil.CreateBlock(ctx, dir, series, 10, 1000, 2000, labels.FromStrings("cluster", "a", "region", "r1"), 0, metadata.NoneFunc, nil)
 	testutil.Ok(t, err)
 	testutil.Ok(t, block.Upload(ctx, logger, bkt, filepath.Join(dir, id2.String()), metadata.NoneFunc))
 
-	id3, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "b", "region", "r1"), 0, metadata.NoneFunc)
+	id3, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "b", "region", "r1"), 0, metadata.NoneFunc, nil)
 	testutil.Ok(t, err)
 	testutil.Ok(t, block.Upload(ctx, logger, bkt, filepath.Join(dir, id3.String()), metadata.NoneFunc))
 
-	id4, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "a", "region", "r2"), 0, metadata.NoneFunc)
+	id4, err := e2eutil.CreateBlock(ctx, dir, series, 10, 0, 1000, labels.FromStrings("cluster", "a", "region", "r2"), 0, metadata.NoneFunc, nil)
 	testutil.Ok(t, err)
 	testutil.Ok(t, block.Upload(ctx, logger, bkt, filepath.Join(dir, id4.String()), metadata.NoneFunc))
 
@@ -3488,6 +3488,62 @@ func TestPostingGroupMerge(t *testing.T) {
 			}
 			res := tc.group1.mergeKeys(tc.group2)
 			testutil.Equals(t, tc.expected, res)
+		})
+	}
+}
+
+func TestToPostingGroup(t *testing.T) {
+	ctx := context.Background()
+	for _, tc := range []struct {
+		name                  string
+		vals                  []string
+		matcher               *labels.Matcher
+		addAll                bool
+		addKeys               []string
+		removeKeys            []string
+		numberLabelValueCalls int
+	}{
+		{
+			name:                  "regexp .*",
+			matcher:               labels.MustNewMatcher(labels.MatchRegexp, labels.MetricName, ".*"),
+			addAll:                true,
+			numberLabelValueCalls: 0,
+		},
+		{
+			name:                  "not regexp .*",
+			matcher:               labels.MustNewMatcher(labels.MatchNotRegexp, labels.MetricName, ".*"),
+			addAll:                false,
+			numberLabelValueCalls: 0,
+		},
+		{
+			name:                  "regexp .+",
+			matcher:               labels.MustNewMatcher(labels.MatchRegexp, labels.MetricName, ".+"),
+			addAll:                false,
+			addKeys:               []string{"foo"},
+			vals:                  []string{"foo"},
+			numberLabelValueCalls: 1,
+		},
+		{
+			name:                  "not regexp .+",
+			matcher:               labels.MustNewMatcher(labels.MatchNotRegexp, labels.MetricName, ".+"),
+			addAll:                true,
+			removeKeys:            []string{"foo"},
+			vals:                  []string{"foo"},
+			numberLabelValueCalls: 1,
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			calls := 0
+			lvalsFn := func(name string) ([]string, error) {
+				calls++
+				return tc.vals, nil
+			}
+			pg, _, err := toPostingGroup(ctx, lvalsFn, tc.matcher)
+			testutil.Ok(t, err)
+			testutil.Equals(t, tc.addAll, pg.addAll)
+			testutil.Equals(t, tc.addKeys, pg.addKeys)
+			testutil.Equals(t, tc.removeKeys, pg.removeKeys)
+			testutil.Equals(t, tc.numberLabelValueCalls, calls)
 		})
 	}
 }